cant do new tweets

+++ app/Http/Controllers/Tweet/DeleteController.php
@@ -3,13 +3,18 @@
 namespace App\Http\Controllers\Tweet;
 
 use App\Http\Controllers\Controller;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
+
 use Illuminate\Http\Request;
 use App\Models\Tweet;
-
+use App\Services\TweetsService;
 class DeleteController extends Controller
 {
     //
-    public function e(Request $request){
+    public function e(Request $request, TweetService $tweetService){
+        if(!$tweetService->checkOwnTweet($request->user()=>id, $request-id())){
+            throw new AccessDeniedHttpException();
+        }
         $tweetId = (int) $request->route('tweetId');
         $tweets = Tweet::where('id', $tweetId)->firstOrFail();
         $tweets->delete();
